fix: npx changeset
