+++ actions/check-changeset/package.json
@@ -22,10 +22,10 @@
     "dev": "concurrently \"pnpm run watch-compile\" \"pnpm run watch-dev\"",
     "lint": "eslint --ext .ts src",
     "lint:fix": "eslint --fix --ext .ts src",
-    "test": "npx jest --runInBand --detectOpenHandles --forceExit",
-    "test:file": "npx jest --detectOpenHandles --forceExit --maxWorkers 1 --testNamePattern ",
-    "test:file:watch": "npx jest --watch --detectOpenHandles --forceExit --maxWorkers 1 --onlyChanged --testNamePattern ",
-    "test:report": "npx jest --runInBand --detectOpenHandles --forceExit --coverage",
+    "test": "./scripts/run_test.sh npx jest --runInBand --detectOpenHandles --forceExit",
+    "test:file": "./scripts/run_test.sh npx jest --detectOpenHandles --forceExit --maxWorkers 1 --testNamePattern ",
+    "test:file:watch": "./scripts/run_test.sh npx jest --watch --detectOpenHandles --forceExit --maxWorkers 1 --onlyChanged --testNamePattern ",
+    "test:report": "./scripts/run_test.sh npx jest --runInBand --detectOpenHandles --forceExit --coverage",
     "start": "node ./dist/index.js",
     "type-check": "tsc --noEmit --project tsconfig.json",
     "watch-compile": "swc src -w --out-dir dist",
