+++ pnpm-workspace.yaml
@@ -1,5 +1,5 @@
 packages:
   - "apps/**/*"
   - "packages/**/*"
-  - "builders/**/*"
+  - "actions/**/*"
   - "!**/dist/**"
